fix: remove custom icon handler for outfits
I believe none of the outfits have a `SmallPreviewImage` and `LargePreviewImage` anymore. We are able to get the icons from the default `IconHandler` itself. so removed it.

`CID_Bean` also doesn't have any icon or anything. So added it to `DisAllowNames` although I am not sure what a 'Bean' outfit is.